REM-21 Refactor TorrentItem component structure and styles
- Moved TorrentItem and its related components (TorrentItemActions, TorrentItemHeader, TorrentItemProgress, TorrentItemStats) into a dedicated directory.
- Updated import paths accordingly.
- Consolidated styles into respective component directories.
- Removed unused styles and components to streamline the codebase.
- Ensured all components maintain their functionality after restructuring.
