+++ frontend/src/components/TorrentItem/TorrentItem.module.css
@@ -1,13 +1,3 @@
-/* Базовые стили компонентов */
-.torrentItem {
-  background: none;
-  border: none;
-  width: 100%;
-  padding: 0;
-  text-align: left;
-  cursor: pointer;
-}
-
 /* Карточка торрента */
 .card {
   margin-bottom: 8px;
@@ -51,45 +41,6 @@
   min-width: 0;
 }
 
-.textEllipsis {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  flex: 1;
-}
-
-.progressWrapper {
-  margin-bottom: 12px;
-}
-
-/* Иконки */
-.downloadIcon { color: var(--blue-9); }
-.uploadIcon { color: var(--grass-9); }
-
-.speedIcon {
-  width: 14px;
-  height: 14px;
-  display: inline-block;
-  vertical-align: middle;
-  margin: 0 2px;
-}
-
-.speedIcon.slow { color: var(--amber-9); }
-
-.speed {
-  display: inline-flex;
-  align-items: center;
-  margin-right: 12px;
-}
-
-/* Действия */
-.actions {
-  display: flex;
-  gap: 8px;
-  margin-left: 16px;
-  -webkit-app-region: no-drag;
-}
-
 /* Темная тема */
 @media (prefers-color-scheme: dark) {
   .card {
@@ -108,7 +59,4 @@
   .cardQueued::before,
   .cardQueuedCheck::before,
   .cardQueuedDownload::before { background-color: var(--purple-8); }
-
-  .downloadIcon { color: var(--blue-8); }
-  .uploadIcon { color: var(--grass-8); }
 }
