+++ frontend/src/components/TorrentItem/TorrentItemActions/index.ts
@@ -1,0 +1,1 @@
+export { TorrentItemActions } from "./TorrentItemActions";
