+++ frontend/src/components/TorrentItem/TorrentItemHeader/index.ts
@@ -1,0 +1,1 @@
+export { TorrentItemHeader } from "./TorrentItemHeader";
