+++ frontend/src/components/TorrentItem/TorrentItemProgress/TorrentItemProgress.module.css
@@ -1,0 +1,4 @@
+/* Стили для прогресс-бара */
+.progressWrapper {
+  margin-bottom: 12px;
+}
