+++ frontend/src/components/TorrentItem/TorrentItemProgress/index.ts
@@ -1,0 +1,1 @@
+export { TorrentItemProgress } from "./TorrentItemProgress";
