+++ frontend/src/components/TorrentItem/TorrentItemStats/index.ts
@@ -1,0 +1,1 @@
+export { TorrentItemStats } from "./TorrentItemStats";
