+++ frontend/src/components/TorrentItem/index.ts
@@ -1,0 +1,1 @@
+export { TorrentItem } from "./TorrentItem";
